uint32_t get_cpu_count() : support >255 threads

+++ linux/cpu.cc
@@ -23,7 +23,7 @@
 #include "cpu.h"
 #include "luts.h"
 
-uint8_t get_cpu_count()
+uint32_t get_cpu_count()
 {
   return sysconf( _SC_NPROCESSORS_ONLN );
 }
